dev_requirements.txt
# Copyright 2020 Thomas Rogers
# SPDX-License-Identifier: Apache-2.0

-r ./requirements.txt
opencv-python==4.4.0.46
antlr4-python3-runtime==4.9.1
